selftests: mptcp: fail if not enough SYN/3rd ACK
Git-commit: 5f88117f256507fc2d146627a3e39bb0cc282a11
Message-Id: <5f88117f256507fc2d146627a3e39bb0cc282a11.1637074297.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc1

If we receive less MPCapable SYN or 3rd ACK than expected, we now mark
the test as failed.

On the other hand, if we receive more, we keep the warning but we add a
hint that it is probably due to retransmissions and that's why we don't
mark the test as failed.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/148
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc1-selftests-mptcp-fail-if-not-enough-SYN-3rd-ACK.patch