mptcp: schedule worker when subflow is closed
Git-commit: 40947e13997a1cba4e875893ca6e5d5e61a0689d
Message-Id: <40947e13997a1cba4e875893ca6e5d5e61a0689d.1637074250.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc1

When remote side closes a subflow we should schedule the worker to
dispose of the subflow in a timely manner.

Otherwise, SF_CLOSED event won't be generated until the mptcp
socket itself is closing or local side is closing another subflow.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc1-mptcp-schedule-worker-when-subflow-is-closed.patch